test: separate comment modules
