+++ packages/bundle-source/demo/comments/block-closer.js
@@ -1,0 +1,3 @@
+/* global success */
+// */ [not executed]
+success();
