+++ packages/bundle-source/demo/comments/block-opener.js
@@ -1,0 +1,4 @@
+/* global success */
+// /*
+success();
+// */
