+++ packages/bundle-source/demo/comments/trailing-comment.js
@@ -11,6 +11,3 @@ export function buildRootObject() {
   });
 }
 // comment
-// */ fooBarBaz!
-
-// /*
